Big grammar cleanup
Changes mainly relate to turning off backtracking #54, as that has
hidden all warnings coming from the parser generator, thus effectively
hiding serious issues in the grammar.

In general, the metamodel and parser became more lax, with the addition
of some validation rules. Additional rules might be needed in the
future, as some cases needs to be handled (after identification).

Furthermore, this change should have solved #238 as well, but that needs
further testing.